requirements.txt
aiosqlite>=0.19.0
asyncpg>=0.31.0
fastapi>=0.128.0
greenlet>=3.3.1
pydantic-settings>=2.0.0
httpx>=0.27.0
pytest>=9.0.2
pytest-asyncio>=1.3.0
requests>=2.32.5
sqlalchemy>=2.0.46
uvicorn>=0.40.0
black>=26.1.0
pytest-cov>=7.0.0
